float/double to float32/float64 to be more consistant

+++ cgogn/geometry/types/plane_3d.cpp
@@ -34,8 +34,8 @@ namespace geometry
 
 template class CGOGN_GEOMETRY_API Plane3D<Eigen::Vector3d>;
 template class CGOGN_GEOMETRY_API Plane3D<Eigen::Vector3f>;
-template class CGOGN_GEOMETRY_API Plane3D<Vec_T<std::array<double,3>>>;
-template class CGOGN_GEOMETRY_API Plane3D<Vec_T<std::array<float,3>>>;
+template class CGOGN_GEOMETRY_API Plane3D<Vec_T<std::array<float32,3>>>;
+template class CGOGN_GEOMETRY_API Plane3D<Vec_T<std::array<float64,3>>>;
 
 } // namespace geometry
